已完成： 完成整个后端。准备开发安卓端 old: 已将CommentTeacher和TeachInfo改造为多对多的关系（涉及到删除，添加时的中间表处理，使用原生查询时要注意service层的事务，以及多对多关联的设置，inner join fetch和left join fetch）
代办：
js前台校验，核对下。
登陆界面弄好看点

+++ src/main/java/Evaluation/dao/AdminDao.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminDao extends JpaRepository<Admin,Long> {
+public interface AdminDao extends JpaRepository<Admin,Integer> {
     Admin findByUsername(String username);
 }
